Use environment variable and remove mock data
