+++ src/apiCalls.ts
@@ -11,7 +11,7 @@ const getAnimalInfo = async(endpoints: string) => {
   console.log('api key', process.env.REACT_APP_API_KEY)
   const response = await fetch(`https://api.api-ninjas.com/v1/${endpoints}`, {
     headers: {
-      "X-Api-Key": "VOXndYvMMax5nCDF8Jz4tQ==blyuYdVC9DZcGh5Z"
+      "X-Api-Key": process.env.REACT_APP_API_KEY!
     }
   })
   console.log('res', response)
