+++ src/components/Results/BackupResults.ts
@@ -1,0 +1,46 @@
+import { Dog, Cat } from "../../types";
+
+export const Abyssinian: Cat = {
+  "length": "12 to 16 inches",
+  "origin": "Southeast Asia",
+  "image_link": "https://api-ninjas.com/images/cats/abyssinian.jpg",
+  "family_friendly": 3,
+  "shedding": 3,
+  "general_health": 2,
+  "playfulness": 5,
+  "children_friendly": 5,
+  "grooming": 3,
+  "other_pets_friendly": 5,
+  "min_weight": 6,
+  "max_weight": 10,
+  "min_life_expectancy": 9,
+  "max_life_expectancy": 15,
+  "name": "Abyssinian"
+}
+
+export const GoldenRetriver: Dog = {
+  "image_link": "https://api-ninjas.com/images/dogs/golden_retriever.jpg",
+  "good_with_children": 5,
+  "good_with_other_dogs": 5,
+  "shedding": 4,
+  "grooming": 2,
+  "drooling": 2,
+  "coat_length": 1,
+  "good_with_strangers": 5,
+  "playfulness": 4,
+  "protectiveness": 3,
+  "trainability": 5,
+  "energy": 3,
+  "barking": 1,
+  "min_life_expectancy": 10,
+  "max_life_expectancy": 12,
+  "max_height_male": 24,
+  "max_height_female": 24,
+  "max_weight_male": 75,
+  "max_weight_female": 65,
+  "min_height_male": 23,
+  "min_height_female": 23,
+  "min_weight_male": 65,
+  "min_weight_female": 55,
+  "name": "Golden Retriever"
+}
