--- src/components/Results/MockData.ts
@@ -1,1799 +1,0 @@
-import { Indexable, Dog, Cat } from "../../types";
-
-const dogs: Dog[] = [
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/affenpinscher.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 3,
-      "drooling": 1,
-      "coat_length": 2,
-      "good_with_strangers": 5,
-      "playfulness": 3,
-      "protectiveness": 3,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 11.5,
-      "max_height_female": 11.5,
-      "max_weight_male": 10.0,
-      "max_weight_female": 10.0,
-      "min_height_male": 9.0,
-      "min_height_female": 9.0,
-      "min_weight_male": 7.0,
-      "min_weight_female": 7.0,
-      "name": "Affenpinscher"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/airedale_terrier.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 1,
-      "grooming": 3,
-      "drooling": 1,
-      "coat_length": 2,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 5,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 11.0,
-      "max_life_expectancy": 14.0,
-      "max_height_male": 23.0,
-      "max_height_female": 23.0,
-      "max_weight_male": 70.0,
-      "max_weight_female": 70.0,
-      "min_height_male": 23.0,
-      "min_height_female": 23.0,
-      "min_weight_male": 50.0,
-      "min_weight_female": 50.0,
-      "name": "Airedale Terrier"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/american_hairless_terrier.jpg",
-      "good_with_children": 5,
-      "good_with_other_dogs": 3,
-      "shedding": 1,
-      "grooming": 1,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 3,
-      "trainability": 5,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 14.0,
-      "max_life_expectancy": 16.0,
-      "max_height_male": 16.0,
-      "max_height_female": 16.0,
-      "max_weight_male": 16.0,
-      "max_weight_female": 16.0,
-      "min_height_male": 12.0,
-      "min_height_female": 12.0,
-      "min_weight_male": 12.0,
-      "min_weight_female": 12.0,
-      "name": "American Hairless Terrier"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/american_staffordshire_terrier.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 2,
-      "grooming": 1,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 4,
-      "playfulness": 3,
-      "protectiveness": 5,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 16.0,
-      "max_height_male": 19.0,
-      "max_height_female": 19.0,
-      "max_weight_male": 70.0,
-      "max_weight_female": 55.0,
-      "min_height_male": 18.0,
-      "min_height_female": 18.0,
-      "min_weight_male": 55.0,
-      "min_weight_female": 40.0,
-      "name": "American Staffordshire Terrier"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/american_water_spaniel.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 1,
-      "grooming": 3,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 3,
-      "trainability": 5,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 14.0,
-      "max_height_male": 18.0,
-      "max_height_female": 18.0,
-      "max_weight_male": 45.0,
-      "max_weight_female": 40.0,
-      "min_height_male": 15.0,
-      "min_height_female": 15.0,
-      "min_weight_male": 30.0,
-      "min_weight_female": 25.0,
-      "name": "American Water Spaniel"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/anatolian_shepherd_dog.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 2,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 1,
-      "playfulness": 3,
-      "protectiveness": 5,
-      "trainability": 2,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 11.0,
-      "max_life_expectancy": 13.0,
-      "max_height_male": 29.0,
-      "max_height_female": 29.0,
-      "max_weight_male": 150.0,
-      "max_weight_female": 120.0,
-      "min_height_male": 29.0,
-      "min_height_female": 29.0,
-      "min_weight_male": 110.0,
-      "min_weight_female": 80.0,
-      "name": "Anatolian Shepherd Dog"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/appenzeller_sennenhund.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 2,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 3,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 4,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 22.0,
-      "max_height_female": 22.0,
-      "max_weight_male": 70.0,
-      "max_weight_female": 70.0,
-      "min_height_male": 19.0,
-      "min_height_female": 19.0,
-      "min_weight_male": 48.0,
-      "min_weight_female": 48.0,
-      "name": "Appenzeller Sennenhund"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/azawakh.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 2,
-      "grooming": 2,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 1,
-      "playfulness": 3,
-      "protectiveness": 3,
-      "trainability": 2,
-      "energy": 3,
-      "barking": 1,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 29.0,
-      "max_height_female": 29.0,
-      "max_weight_male": 55.0,
-      "max_weight_female": 44.0,
-      "min_height_male": 25.0,
-      "min_height_female": 25.0,
-      "min_weight_male": 44.0,
-      "min_weight_female": 33.0,
-      "name": "Azawakh"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/barbet.jpg",
-      "good_with_children": 5,
-      "good_with_other_dogs": 5,
-      "shedding": 1,
-      "grooming": 3,
-      "drooling": 1,
-      "coat_length": 2,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 3,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 14.0,
-      "max_height_male": 24.5,
-      "max_height_female": 24.5,
-      "max_weight_male": 65.0,
-      "max_weight_female": 65.0,
-      "min_height_male": 19.0,
-      "min_height_female": 19.0,
-      "min_weight_male": 35.0,
-      "min_weight_female": 35.0,
-      "name": "Barbet"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/bergamasco_sheepdog.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 1,
-      "grooming": 1,
-      "drooling": 2,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 4,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 1,
-      "min_life_expectancy": 13.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 23.5,
-      "max_height_female": 23.5,
-      "max_weight_male": 84.0,
-      "max_weight_female": 71.0,
-      "min_height_male": 23.5,
-      "min_height_female": 23.5,
-      "min_weight_male": 70.0,
-      "min_weight_female": 57.0,
-      "name": "Bergamasco Sheepdog"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/biewer_terrier.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 5,
-      "shedding": 1,
-      "grooming": 3,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 4,
-      "protectiveness": 2,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 16.0,
-      "max_life_expectancy": 16.0,
-      "max_height_male": 11.0,
-      "max_height_female": 11.0,
-      "max_weight_male": 8.0,
-      "max_weight_female": 8.0,
-      "min_height_male": 7.0,
-      "min_height_female": 7.0,
-      "min_weight_male": 4.0,
-      "min_weight_female": 4.0,
-      "name": "Biewer Terrier"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/black_and_tan_coonhound.jpg",
-      "good_with_children": 5,
-      "good_with_other_dogs": 5,
-      "shedding": 3,
-      "grooming": 2,
-      "drooling": 3,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 2,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 4,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 12.0,
-      "max_height_male": 27.0,
-      "max_height_female": 27.0,
-      "max_weight_male": 110.0,
-      "max_weight_female": 110.0,
-      "min_height_male": 25.0,
-      "min_height_female": 25.0,
-      "min_weight_male": 65.0,
-      "min_weight_female": 65.0,
-      "name": "Black and Tan Coonhound"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/bloodhound.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 2,
-      "drooling": 5,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 2,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 5,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 12.0,
-      "max_height_male": 27.0,
-      "max_height_female": 27.0,
-      "max_weight_male": 110.0,
-      "max_weight_female": 100.0,
-      "min_height_male": 25.0,
-      "min_height_female": 25.0,
-      "min_weight_male": 90.0,
-      "min_weight_female": 80.0,
-      "name": "Bloodhound"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/boerboel.jpg",
-      "good_with_children": 4,
-      "good_with_other_dogs": 2,
-      "shedding": 3,
-      "grooming": 2,
-      "drooling": 3,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 5,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 11.0,
-      "max_height_male": 27.0,
-      "max_height_female": 27.0,
-      "max_weight_male": 200.0,
-      "max_weight_female": 200.0,
-      "min_height_male": 24.0,
-      "min_height_female": 24.0,
-      "min_weight_male": 150.0,
-      "min_weight_female": 150.0,
-      "name": "Boerboel"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/bohemian_shepherd.jpg",
-      "good_with_children": 5,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 2,
-      "drooling": 2,
-      "coat_length": 1,
-      "good_with_strangers": 4,
-      "playfulness": 3,
-      "protectiveness": 3,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 21.7,
-      "max_height_female": 21.7,
-      "max_weight_male": 60.0,
-      "max_weight_female": 53.0,
-      "min_height_male": 20.5,
-      "min_height_female": 20.5,
-      "min_weight_male": 41.0,
-      "min_weight_female": 37.0,
-      "name": "Bohemian Shepherd"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/border_terrier.jpg",
-      "good_with_children": 5,
-      "good_with_other_dogs": 3,
-      "shedding": 2,
-      "grooming": 2,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 4,
-      "playfulness": 4,
-      "protectiveness": 3,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 15.0,
-      "max_height_female": 15.0,
-      "max_weight_male": 15.5,
-      "max_weight_female": 14.0,
-      "min_height_male": 12.0,
-      "min_height_female": 12.0,
-      "min_weight_male": 13.0,
-      "min_weight_female": 11.5,
-      "name": "Border Terrier"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/briard.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 1,
-      "grooming": 4,
-      "drooling": 2,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 4,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 1,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 12.0,
-      "max_height_male": 27.0,
-      "max_height_female": 27.0,
-      "max_weight_male": 100.0,
-      "max_weight_female": 100.0,
-      "min_height_male": 23.0,
-      "min_height_female": 23.0,
-      "min_weight_male": 55.0,
-      "min_weight_female": 55.0,
-      "name": "Briard"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/brussels_griffon.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 3,
-      "drooling": 1,
-      "coat_length": 2,
-      "good_with_strangers": 4,
-      "playfulness": 4,
-      "protectiveness": 3,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 4,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 10.0,
-      "max_height_female": 10.0,
-      "max_weight_male": 10.0,
-      "max_weight_female": 10.0,
-      "min_height_male": 7.0,
-      "min_height_female": 7.0,
-      "min_weight_male": 8.0,
-      "min_weight_female": 8.0,
-      "name": "Brussels Griffon"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/bulldog.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 3,
-      "drooling": 3,
-      "coat_length": 1,
-      "good_with_strangers": 4,
-      "playfulness": 4,
-      "protectiveness": 3,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 2,
-      "min_life_expectancy": 8.0,
-      "max_life_expectancy": 10.0,
-      "max_height_male": 15.0,
-      "max_height_female": 15.0,
-      "max_weight_male": 50.0,
-      "max_weight_female": 40.0,
-      "min_height_male": 14.0,
-      "min_height_female": 14.0,
-      "min_weight_male": 50.0,
-      "min_weight_female": 40.0,
-      "name": "Bulldog"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/cairn_terrier.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 2,
-      "grooming": 2,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 4,
-      "protectiveness": 4,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 4,
-      "min_life_expectancy": 13.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 10.0,
-      "max_height_female": 10.0,
-      "max_weight_male": 14.0,
-      "max_weight_female": 13.0,
-      "min_height_male": 10.0,
-      "min_height_female": 10.0,
-      "min_weight_male": 14.0,
-      "min_weight_female": 13.0,
-      "name": "Cairn Terrier"
-  }
-]
-
-const dogs3: Dog[]= [{
-      "image_link": "https://api-ninjas.com/images/dogs/bloodhound.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 2,
-      "drooling": 5,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 2,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 5,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 12.0,
-      "max_height_male": 27.0,
-      "max_height_female": 27.0,
-      "max_weight_male": 110.0,
-      "max_weight_female": 100.0,
-      "min_height_male": 25.0,
-      "min_height_female": 25.0,
-      "min_weight_male": 90.0,
-      "min_weight_female": 80.0,
-      "name": "Bloodhound"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/boerboel.jpg",
-      "good_with_children": 4,
-      "good_with_other_dogs": 2,
-      "shedding": 3,
-      "grooming": 2,
-      "drooling": 3,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 5,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 11.0,
-      "max_height_male": 27.0,
-      "max_height_female": 27.0,
-      "max_weight_male": 200.0,
-      "max_weight_female": 200.0,
-      "min_height_male": 24.0,
-      "min_height_female": 24.0,
-      "min_weight_male": 150.0,
-      "min_weight_female": 150.0,
-      "name": "Boerboel"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/bohemian_shepherd.jpg",
-      "good_with_children": 5,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 2,
-      "drooling": 2,
-      "coat_length": 1,
-      "good_with_strangers": 4,
-      "playfulness": 3,
-      "protectiveness": 3,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 21.7,
-      "max_height_female": 21.7,
-      "max_weight_male": 60.0,
-      "max_weight_female": 53.0,
-      "min_height_male": 20.5,
-      "min_height_female": 20.5,
-      "min_weight_male": 41.0,
-      "min_weight_female": 37.0,
-      "name": "Bohemian Shepherd"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/border_terrier.jpg",
-      "good_with_children": 5,
-      "good_with_other_dogs": 3,
-      "shedding": 2,
-      "grooming": 2,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 4,
-      "playfulness": 4,
-      "protectiveness": 3,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 15.0,
-      "max_height_female": 15.0,
-      "max_weight_male": 15.5,
-      "max_weight_female": 14.0,
-      "min_height_male": 12.0,
-      "min_height_female": 12.0,
-      "min_weight_male": 13.0,
-      "min_weight_female": 11.5,
-      "name": "Border Terrier"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/briard.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 1,
-      "grooming": 4,
-      "drooling": 2,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 4,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 1,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 12.0,
-      "max_height_male": 27.0,
-      "max_height_female": 27.0,
-      "max_weight_male": 100.0,
-      "max_weight_female": 100.0,
-      "min_height_male": 23.0,
-      "min_height_female": 23.0,
-      "min_weight_male": 55.0,
-      "min_weight_female": 55.0,
-      "name": "Briard"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/brussels_griffon.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 3,
-      "drooling": 1,
-      "coat_length": 2,
-      "good_with_strangers": 4,
-      "playfulness": 4,
-      "protectiveness": 3,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 4,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 10.0,
-      "max_height_female": 10.0,
-      "max_weight_male": 10.0,
-      "max_weight_female": 10.0,
-      "min_height_male": 7.0,
-      "min_height_female": 7.0,
-      "min_weight_male": 8.0,
-      "min_weight_female": 8.0,
-      "name": "Brussels Griffon"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/bulldog.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 3,
-      "drooling": 3,
-      "coat_length": 1,
-      "good_with_strangers": 4,
-      "playfulness": 4,
-      "protectiveness": 3,
-      "trainability": 4,
-      "energy": 3,
-      "barking": 2,
-      "min_life_expectancy": 8.0,
-      "max_life_expectancy": 10.0,
-      "max_height_male": 15.0,
-      "max_height_female": 15.0,
-      "max_weight_male": 50.0,
-      "max_weight_female": 40.0,
-      "min_height_male": 14.0,
-      "min_height_female": 14.0,
-      "min_weight_male": 50.0,
-      "min_weight_female": 40.0,
-      "name": "Bulldog"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/cairn_terrier.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 2,
-      "grooming": 2,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 4,
-      "protectiveness": 4,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 4,
-      "min_life_expectancy": 13.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 10.0,
-      "max_height_female": 10.0,
-      "max_weight_male": 14.0,
-      "max_weight_female": 13.0,
-      "min_height_male": 10.0,
-      "min_height_female": 10.0,
-      "min_weight_male": 14.0,
-      "min_weight_female": 13.0,
-      "name": "Cairn Terrier"
-  }
-]
-
-const dogs2: Dog[] = [
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/affenpinscher.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 3,
-      "grooming": 3,
-      "drooling": 1,
-      "coat_length": 2,
-      "good_with_strangers": 5,
-      "playfulness": 3,
-      "protectiveness": 3,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "max_height_male": 11.5,
-      "max_height_female": 11.5,
-      "max_weight_male": 10.0,
-      "max_weight_female": 10.0,
-      "min_height_male": 9.0,
-      "min_height_female": 9.0,
-      "min_weight_male": 7.0,
-      "min_weight_female": 7.0,
-      "name": "Affenpinscher"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/airedale_terrier.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 1,
-      "grooming": 3,
-      "drooling": 1,
-      "coat_length": 2,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 5,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 11.0,
-      "max_life_expectancy": 14.0,
-      "max_height_male": 23.0,
-      "max_height_female": 23.0,
-      "max_weight_male": 70.0,
-      "max_weight_female": 70.0,
-      "min_height_male": 23.0,
-      "min_height_female": 23.0,
-      "min_weight_male": 50.0,
-      "min_weight_female": 50.0,
-      "name": "Airedale Terrier"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/american_hairless_terrier.jpg",
-      "good_with_children": 5,
-      "good_with_other_dogs": 3,
-      "shedding": 1,
-      "grooming": 1,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 3,
-      "playfulness": 3,
-      "protectiveness": 3,
-      "trainability": 5,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 14.0,
-      "max_life_expectancy": 16.0,
-      "max_height_male": 16.0,
-      "max_height_female": 16.0,
-      "max_weight_male": 16.0,
-      "max_weight_female": 16.0,
-      "min_height_male": 12.0,
-      "min_height_female": 12.0,
-      "min_weight_male": 12.0,
-      "min_weight_female": 12.0,
-      "name": "American Hairless Terrier"
-  },
-  {
-      "image_link": "https://api-ninjas.com/images/dogs/american_staffordshire_terrier.jpg",
-      "good_with_children": 3,
-      "good_with_other_dogs": 3,
-      "shedding": 2,
-      "grooming": 1,
-      "drooling": 1,
-      "coat_length": 1,
-      "good_with_strangers": 4,
-      "playfulness": 3,
-      "protectiveness": 5,
-      "trainability": 3,
-      "energy": 3,
-      "barking": 3,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 16.0,
-      "max_height_male": 19.0,
-      "max_height_female": 19.0,
-      "max_weight_male": 70.0,
-      "max_weight_female": 55.0,
-      "min_height_male": 18.0,
-      "min_height_female": 18.0,
-      "min_weight_male": 55.0,
-      "min_weight_female": 40.0,
-      "name": "American Staffordshire Terrier"
-  }
-]
-
-export const cats: Cat[] = [
-  {
-      "length": "12 to 16 inches",
-      "origin": "Southeast Asia",
-      "image_link": "https://api-ninjas.com/images/cats/abyssinian.jpg",
-      "family_friendly": 3,
-      "shedding": 3,
-      "general_health": 2,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 3,
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 15.0,
-      "name": "Abyssinian"
-  },
-  {
-      "length": "Medium",
-      "origin": "Greece",
-      "image_link": "https://api-ninjas.com/images/cats/aegean.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 4,
-      "other_pets_friendly": 3,
-      "min_weight": 7.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 10.0,
-      "name": "Aegean"
-  },
-  {
-      "length": "12 to 15 inches",
-      "origin": "United States",
-      "image_link": "https://api-ninjas.com/images/cats/american_shorthair.jpg",
-      "family_friendly": 3,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 2,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 3,
-      "min_weight": 7.0,
-      "max_weight": 12.0,
-      "min_life_expectancy": 15.0,
-      "max_life_expectancy": 20.0,
-      "name": "American Shorthair"
-  },
-  {
-      "length": "Medium to large",
-      "origin": "United States",
-      "image_link": "https://api-ninjas.com/images/cats/american_wirehair.jpg",
-      "family_friendly": 4,
-      "shedding": 3,
-      "general_health": 5,
-      "playfulness": 3,
-      "children_friendly": 4,
-      "grooming": 5,
-      "other_pets_friendly": 3,
-      "min_weight": 8.0,
-      "max_weight": 15.0,
-      "min_life_expectancy": 14.0,
-      "max_life_expectancy": 18.0,
-      "name": "American Wirehair"
-  },
-  {
-      "length": "Up to 18 inches, not including tail",
-      "origin": "United States and Thailand",
-      "image_link": "https://api-ninjas.com/images/cats/balinese.jpg",
-      "family_friendly": 4,
-      "shedding": 3,
-      "general_health": 2,
-      "playfulness": 4,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 15.0,
-      "name": "Balinese"
-  },
-  {
-      "length": "15 to 18 inches",
-      "origin": "Unknown",
-      "image_link": "https://api-ninjas.com/images/cats/birman.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 2,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 2,
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 12.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 16.0,
-      "name": "Birman"
-  },
-  {
-      "length": "22 to 25 inches, not including tail",
-      "origin": "Great Britain",
-      "image_link": "https://api-ninjas.com/images/cats/british_shorthair.jpg",
-      "family_friendly": 3,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 2,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 5,
-      "min_weight": 7.0,
-      "max_weight": 17.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 17.0,
-      "name": "British Shorthair"
-  },
-  {
-      "length": "Medium",
-      "origin": "England",
-      "image_link": "https://api-ninjas.com/images/cats/burmilla.jpg",
-      "family_friendly": 4,
-      "shedding": 3,
-      "general_health": 3,
-      "playfulness": 5,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 6.5,
-      "max_weight": 13.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 15.0,
-      "name": "Burmilla"
-  },
-  {
-      "length": "Medium",
-      "origin": "United States",
-      "image_link": "https://api-ninjas.com/images/cats/chantilly-tiffany.jpg",
-      "family_friendly": 4,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 4,
-      "children_friendly": 4,
-      "grooming": 2,
-      "other_pets_friendly": 3,
-      "min_weight": 6.0,
-      "max_weight": 12.0,
-      "min_life_expectancy": 7.0,
-      "max_life_expectancy": 16.0,
-      "name": "Chantilly-Tiffany"
-  },
-  {
-      "length": "Medium to large",
-      "origin": "China",
-      "image_link": "https://api-ninjas.com/images/cats/chinese_li_hua.jpg",
-      "family_friendly": 3,
-      "shedding": 3,
-      "general_health": 3,
-      "playfulness": 4,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 9.0,
-      "max_weight": 12.0,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 16.0,
-      "name": "Chinese Li Hua"
-  },
-  {
-      "length": "Medium",
-      "origin": "Egypt",
-      "image_link": "https://api-ninjas.com/images/cats/egyptian_mau.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 4,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 3,
-      "min_weight": 6.0,
-      "max_weight": 14.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "name": "Egyptian Mau"
-  },
-  {
-      "length": "Medium",
-      "origin": "United States",
-      "image_link": "https://api-ninjas.com/images/cats/exotic.jpg",
-      "family_friendly": 4,
-      "shedding": 3,
-      "general_health": 2,
-      "playfulness": 4,
-      "children_friendly": 3,
-      "grooming": 2,
-      "other_pets_friendly": 3,
-      "min_weight": 7.0,
-      "max_weight": 12.0,
-      "min_life_expectancy": 8.0,
-      "max_life_expectancy": 15.0,
-      "name": "Exotic"
-  },
-  {
-      "length": "Medium",
-      "origin": "Quebec, Canada",
-      "image_link": "https://api-ninjas.com/images/cats/foldex.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 3,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 14.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "name": "Foldex"
-  },
-  {
-      "length": "12 to 15 inches",
-      "origin": "England",
-      "image_link": "https://api-ninjas.com/images/cats/havana_brown.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 3,
-      "children_friendly": 4,
-      "grooming": 5,
-      "other_pets_friendly": 4,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 15.0,
-      "name": "Havana Brown"
-  },
-  {
-      "length": "Small to medium",
-      "origin": "United States and Canada",
-      "image_link": "https://api-ninjas.com/images/cats/javanese.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 2,
-      "playfulness": 4,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 15.0,
-      "name": "Javanese"
-  },
-  {
-      "length": "14 to 16 inches",
-      "origin": "Isle of Man, UK",
-      "image_link": "https://api-ninjas.com/images/cats/manx.jpg",
-      "family_friendly": 4,
-      "shedding": 3,
-      "general_health": 3,
-      "playfulness": 3,
-      "children_friendly": 2,
-      "grooming": 3,
-      "other_pets_friendly": 2,
-      "min_weight": 8.0,
-      "max_weight": 12.0,
-      "min_life_expectancy": 8.0,
-      "max_life_expectancy": 14.0,
-      "name": "Manx"
-  },
-  {
-      "length": "Small",
-      "origin": "USA",
-      "image_link": "https://api-ninjas.com/images/cats/napoleon.jpg",
-      "family_friendly": 4,
-      "shedding": 3,
-      "general_health": 3,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 2,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 9.0,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 15.0,
-      "name": "Napoleon"
-  },
-  {
-      "length": "13 to 20 inches, not including tail",
-      "origin": "USA",
-      "image_link": "https://api-ninjas.com/images/cats/savannah.jpg",
-      "family_friendly": 4,
-      "shedding": 3,
-      "general_health": 5,
-      "playfulness": 5,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 5,
-      "min_weight": 8.0,
-      "max_weight": 20.0,
-      "min_life_expectancy": 17.0,
-      "max_life_expectancy": 20.0,
-      "name": "Savannah"
-  },
-  {
-      "length": "10 to 12 inches",
-      "origin": "Tayside, Scotland",
-      "image_link": "https://api-ninjas.com/images/cats/scottish_fold.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 3,
-      "playfulness": 3,
-      "children_friendly": 5,
-      "grooming": 2,
-      "other_pets_friendly": 4,
-      "min_weight": 6.0,
-      "max_weight": 13.0,
-      "min_life_expectancy": 11.0,
-      "max_life_expectancy": 14.0,
-      "name": "Scottish Fold"
-  }
-]
-
-export const cats2: Cat[] = [
-  {
-      "length": "12 to 16 inches",
-      "origin": "Southeast Asia",
-      "image_link": "https://api-ninjas.com/images/cats/abyssinian.jpg",
-      "family_friendly": 3,
-      "shedding": 3,
-      "general_health": 2,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 3,
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 15.0,
-      "name": "Abyssinian"
-  },
-  {
-      "length": "17 to 22 inches, not including tail",
-      "origin": "USA",
-      "image_link": "https://api-ninjas.com/images/cats/bengal_cats.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 3,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 4,
-      "other_pets_friendly": 5,
-      "min_weight": 8.0,
-      "max_weight": 17.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 16.0,
-      "name": "Bengal Cats"
-  },
-  {
-      "length": "Medium",
-      "origin": "England",
-      "image_link": "https://api-ninjas.com/images/cats/burmilla.jpg",
-      "family_friendly": 4,
-      "shedding": 3,
-      "general_health": 3,
-      "playfulness": 5,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 6.5,
-      "max_weight": 13.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 15.0,
-      "name": "Burmilla"
-  },
-  {
-      "length": "Small to medium",
-      "origin": "England",
-      "image_link": "https://api-ninjas.com/images/cats/colorpoint_shorthair.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 2,
-      "playfulness": 5,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 17.0,
-      "name": "Colorpoint Shorthair"
-  },
-  {
-      "length": "Small",
-      "origin": "United Kingdom",
-      "image_link": "https://api-ninjas.com/images/cats/cornish_rex.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 3,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 5,
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 11.0,
-      "max_life_expectancy": 15.0,
-      "name": "Cornish Rex"
-  },
-  {
-      "length": "Medium",
-      "origin": "Cyprus",
-      "image_link": "https://api-ninjas.com/images/cats/cyprus.jpg",
-      "family_friendly": 4,
-      "shedding": 2,
-      "general_health": 4,
-      "playfulness": 5,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 8.0,
-      "max_weight": 16.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "name": "Cyprus"
-  },
-  {
-      "length": "Medium",
-      "origin": "Burma and Thailand",
-      "image_link": "https://api-ninjas.com/images/cats/european_burmese.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 1,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 5,
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 15.0,
-      "name": "European Burmese"
-  },
-  {
-      "length": "Medium",
-      "origin": "China and Japan",
-      "image_link": "https://api-ninjas.com/images/cats/japanese_bobtail.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 4,
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 15.0,
-      "name": "Japanese Bobtail"
-  },
-  {
-      "length": "10 to 12 inches",
-      "origin": "Thailand",
-      "image_link": "https://api-ninjas.com/images/cats/khao_manee.jpg",
-      "family_friendly": 4,
-      "shedding": 1,
-      "general_health": 3,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 8.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 12.0,
-      "name": "Khao Manee"
-  },
-  {
-      "length": "Small to medium",
-      "origin": "Oregon, USA",
-      "image_link": "https://api-ninjas.com/images/cats/laperm.jpg",
-      "family_friendly": 4,
-      "shedding": 2,
-      "general_health": 4,
-      "playfulness": 5,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 8.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 15.0,
-      "name": "LaPerm"
-  },
-  {
-      "length": "13 to 20 inches, not including tail",
-      "origin": "USA",
-      "image_link": "https://api-ninjas.com/images/cats/savannah.jpg",
-      "family_friendly": 4,
-      "shedding": 3,
-      "general_health": 5,
-      "playfulness": 5,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 5,
-      "min_weight": 8.0,
-      "max_weight": 20.0,
-      "min_life_expectancy": 17.0,
-      "max_life_expectancy": 20.0,
-      "name": "Savannah"
-  },
-  {
-      "length": "17 to 25 inches",
-      "origin": "Russia",
-      "image_link": "https://api-ninjas.com/images/cats/siberian.jpg",
-      "family_friendly": 5,
-      "shedding": 4,
-      "general_health": 4,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 2,
-      "other_pets_friendly": 4,
-      "min_weight": 8.0,
-      "max_weight": 17.0,
-      "min_life_expectancy": 11.0,
-      "max_life_expectancy": 18.0,
-      "name": "Siberian"
-  },
-  {
-      "length": "Medium to long",
-      "origin": "Philadelphia, USA",
-      "image_link": "https://api-ninjas.com/images/cats/snowshoe.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 5,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 5,
-      "other_pets_friendly": 4,
-      "min_weight": 7.0,
-      "max_weight": 12.0,
-      "min_life_expectancy": 14.0,
-      "max_life_expectancy": 19.0,
-      "name": "Snowshoe"
-  },
-  {
-      "length": "11 to 14 inches",
-      "origin": "United States",
-      "image_link": "https://api-ninjas.com/images/cats/somali.jpg",
-      "family_friendly": 5,
-      "shedding": 4,
-      "general_health": 2,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 3,
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 11.0,
-      "max_life_expectancy": 16.0,
-      "name": "Somali"
-  },
-  {
-      "length": "12 to 15 inches",
-      "origin": "United States and Thailand",
-      "image_link": "https://api-ninjas.com/images/cats/tonkinese.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 2,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 6.0,
-      "max_weight": 12.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 16.0,
-      "name": "Tonkinese"
-  },
-  {
-      "length": "Medium",
-      "origin": "Ankara, Turkey",
-      "image_link": "https://api-ninjas.com/images/cats/turkish_angora.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 3,
-      "playfulness": 5,
-      "children_friendly": 3,
-      "grooming": 4,
-      "other_pets_friendly": 3,
-      "min_weight": 5.0,
-      "max_weight": 9.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 18.0,
-      "name": "Turkish Angora"
-  }
-]
-
-export const cats3: Cat[] = [
-  {
-      "length": "Medium",
-      "origin": "Greece",
-      "image_link": "https://api-ninjas.com/images/cats/aegean.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 4,
-      "other_pets_friendly": 3,
-      "min_weight": 7.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 10.0,
-      "name": "Aegean"
-  },
-  {
-      "length": "Medium",
-      "origin": "California, USA",
-      "image_link": "https://api-ninjas.com/images/cats/american_curl.jpg",
-      "family_friendly": 5,
-      "shedding": 4,
-      "general_health": 5,
-      "playfulness": 4,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 16.0,
-      "name": "American Curl"
-  },
-  {
-      "length": "Medium",
-      "origin": "Great Britain",
-      "image_link": "https://api-ninjas.com/images/cats/asian.jpg",
-      "family_friendly": 5,
-      "shedding": 1,
-      "general_health": 4,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 3,
-      "other_pets_friendly": 4,
-      "min_weight": 10.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 15.0,
-      "max_life_expectancy": 15.0,
-      "name": "Asian"
-  },
-  {
-      "length": "Medium",
-      "origin": "Australia",
-      "image_link": "https://api-ninjas.com/images/cats/australian_mist.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 4,
-      "playfulness": 4,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 3,
-      "min_weight": 8.0,
-      "max_weight": 15.0,
-      "min_life_expectancy": 15.0,
-      "max_life_expectancy": 18.0,
-      "name": "Australian Mist"
-  },
-  {
-      "length": "17 to 22 inches, not including tail",
-      "origin": "USA",
-      "image_link": "https://api-ninjas.com/images/cats/bengal_cats.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 3,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 4,
-      "other_pets_friendly": 5,
-      "min_weight": 8.0,
-      "max_weight": 17.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 16.0,
-      "name": "Bengal Cats"
-  },
-  {
-      "length": "15 to 18 inches",
-      "origin": "Unknown",
-      "image_link": "https://api-ninjas.com/images/cats/birman.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 2,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 2,
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 12.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 16.0,
-      "name": "Birman"
-  },
-  {
-      "length": "15 to 18 inches",
-      "origin": "Burma and Thailand",
-      "image_link": "https://api-ninjas.com/images/cats/burmese.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 2,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 5,
-      "other_pets_friendly": 4,
-      "min_weight": 8.0,
-      "max_weight": 12.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 16.0,
-      "name": "Burmese"
-  },
-  {
-      "length": "Small to medium",
-      "origin": "England",
-      "image_link": "https://api-ninjas.com/images/cats/colorpoint_shorthair.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 2,
-      "playfulness": 5,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 17.0,
-      "name": "Colorpoint Shorthair"
-  },
-  {
-      "length": "Small",
-      "origin": "United Kingdom",
-      "image_link": "https://api-ninjas.com/images/cats/cornish_rex.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 3,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 5,
-
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 11.0,
-      "max_life_expectancy": 15.0,
-      "name": "Cornish Rex"
-  },
-  {
-      "length": "15 to 18 inches",
-      "origin": "Devon, England",
-      "image_link": "https://api-ninjas.com/images/cats/devon_rex.jpg",
-      "family_friendly": 5,
-      "shedding": 1,
-      "general_health": 1,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 5,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 15.0,
-      "name": "Devon Rex"
-  },
-  {
-      "length": "Medium",
-      "origin": "Egypt",
-      "image_link": "https://api-ninjas.com/images/cats/egyptian_mau.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 4,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 3,
-      "min_weight": 6.0,
-      "max_weight": 14.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "name": "Egyptian Mau"
-  },
-  {
-      "length": "Medium",
-      "origin": "Burma and Thailand",
-      "image_link": "https://api-ninjas.com/images/cats/european_burmese.jpg",
-      "family_friendly": 5,
-      "shedding": 2,
-      "general_health": 1,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 5,
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 15.0,
-      "name": "European Burmese"
-  },
-  {
-      "length": "Medium",
-      "origin": "Quebec, Canada",
-      "image_link": "https://api-ninjas.com/images/cats/foldex.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 3,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 14.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "name": "Foldex"
-  },
-  {
-      "length": "12 to 15 inches",
-      "origin": "England",
-      "image_link": "https://api-ninjas.com/images/cats/havana_brown.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 3,
-      "children_friendly": 4,
-      "grooming": 5,
-      "other_pets_friendly": 4,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 15.0,
-      "name": "Havana Brown"
-  },
-  {
-      "length": "Medium",
-      "origin": "China and Japan",
-      "image_link": "https://api-ninjas.com/images/cats/japanese_bobtail.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 4,
-      "playfulness": 5,
-      "children_friendly": 5,
-      "grooming": 4,
-      "other_pets_friendly": 5,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 9.0,
-      "max_life_expectancy": 15.0,
-      "name": "Japanese Bobtail"
-  },
-  {
-      "length": "Small to medium",
-      "origin": "United States and Canada",
-      "image_link": "https://api-ninjas.com/images/cats/javanese.jpg",
-      "family_friendly": 5,
-      "shedding": 3,
-      "general_health": 2,
-      "playfulness": 4,
-      "children_friendly": 4,
-      "grooming": 4,
-      "other_pets_friendly": 4,
-      "min_weight": 5.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 10.0,
-      "max_life_expectancy": 15.0,
-      "name": "Javanese"
-  },
-  {
-      "length": "Medium",
-      "origin": "Russia",
-      "image_link": "https://api-ninjas.com/images/cats/peterbald.jpg",
-      "family_friendly": 5,
-      "shedding": 1,
-      "general_health": 3,
-      "playfulness": 4,
-      "children_friendly": 4,
-      "grooming": 2,
-      "other_pets_friendly": 4,
-      "min_weight": 6.0,
-      "max_weight": 10.0,
-      "min_life_expectancy": 12.0,
-      "max_life_expectancy": 15.0,
-      "name": "Peterbald"
-  },
-  {
-      "length": "20 to 24 inches",
-      "origin": "Washington, USA",
-      "image_link": "https://api-ninjas.com/images/cats/pixie-bob.jpg",
-      "family_friendly": 5,
-      "shedding": 4,
-      "general_health": 5,
-      "playfulness": 4,
-      "children_friendly": 5,
-      "grooming": 2,
-      "other_pets_friendly": 4,
-      "min_weight": 8.0,
-      "max_weight": 17.0,
-      "min_life_expectancy": 13.0,
-      "max_life_expectancy": 15.0,
-      "name": "Pixie-Bob"
-  }
-]
-
-export const Abyssinian: Cat = {
-  "length": "12 to 16 inches",
-  "origin": "Southeast Asia",
-  "image_link": "https://api-ninjas.com/images/cats/abyssinian.jpg",
-  "family_friendly": 3,
-  "shedding": 3,
-  "general_health": 2,
-  "playfulness": 5,
-  "children_friendly": 5,
-  "grooming": 3,
-  "other_pets_friendly": 5,
-  "min_weight": 6,
-  "max_weight": 10,
-  "min_life_expectancy": 9,
-  "max_life_expectancy": 15,
-  "name": "Abyssinian"
-}
-
-export const GoldenRetriver: Dog = {
-  "image_link": "https://api-ninjas.com/images/dogs/golden_retriever.jpg",
-  "good_with_children": 5,
-  "good_with_other_dogs": 5,
-  "shedding": 4,
-  "grooming": 2,
-  "drooling": 2,
-  "coat_length": 1,
-  "good_with_strangers": 5,
-  "playfulness": 4,
-  "protectiveness": 3,
-  "trainability": 5,
-  "energy": 3,
-  "barking": 1,
-  "min_life_expectancy": 10,
-  "max_life_expectancy": 12,
-  "max_height_male": 24,
-  "max_height_female": 24,
-  "max_weight_male": 75,
-  "max_weight_female": 65,
-  "min_height_male": 23,
-  "min_height_female": 23,
-  "min_weight_male": 65,
-  "min_weight_female": 55,
-  "name": "Golden Retriever"
-}
-
-export {dogs, dogs2, dogs3}
