+++ src/components/Results/Results.tsx
@@ -2,11 +2,10 @@ import './Results.css'
 import save from '../../images/save.png'
 import discard from '../../images/discard.png'
 import paw from '../../images/paw-print.png'
-import noResults from '../../images/no-results.png'
 import { Dog, Cat, Indexable, QuizAnswers } from '../../types'
 import { getNames, getRandomAnimal, isCat, isDog } from '../../helpers'
 import { useState, useEffect } from 'react'
-import { GoldenRetriver, Abyssinian, dogs, dogs2, dogs3, cats,  cats2, cats3 } from './MockData'
+import { GoldenRetriver, Abyssinian } from './BackupResults'
 import loadingAnimation from '../../images/loading.gif'
 import { getAnimalInfo } from '../../apiCalls'
